refactor: change utils file name to camel case for consistency
