+++ src/context/cart-and-wishlist-context.js
@@ -1,7 +1,7 @@
 import { createContext, useContext, useReducer } from "react";
 import { cartWishlistInitialState, cartWishlistReducer } from "../reducers";
 import axios from "axios";
-import { isInCart, isInWishlist } from "../utils/cart-and-wishlist-functions";
+import { isInCart, isInWishlist } from "../utils/cartAndWishlistFunctions";
 import { useNavigate ,useLocation} from "react-router-dom";
 import { useAuth } from "./auth-context";
 import { CART_API, WISHLIST_API } from "../constants/apiEndPoints";
