+++ src/utils/cartAndWishlistFunctions.js
