+++ src/utils/throttle.js
@@ -1,0 +1,12 @@
+export const throttle = (func) => {
+  let isDispatchable = true;
+  return function () {
+    if (isDispatchable) {
+      func();
+      isDispatchable = false;
+      setTimeout(() => {
+        isDispatchable = true;
+      }, 1000);
+    }
+  };
+};
